requirements.txt
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
asyncio==3.4.3
click==8.1.7
fastapi==0.110.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
nest-asyncio==1.6.0
packaging==24.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
